Create Problem Statement
In this assignment, I was tasked with tracking the real-time location of a NASA satellite using streaming data from the ISS location API (http://api.open-notify.org/iss-now.json). This API acts as the producer, continuously providing the current longitude and latitude of the satellite in JSON format every time it's called. My job was to write a consumer application that connects to this API, fetches the location data every 5 seconds, and plots the satellite's path on a world map in real time. The consumer was set to run for one hour, collecting position data over time and dynamically updating the graph. Given the high speed of the ISS, the plotted path was expected to span more than half the globe by the end of the hour.